# packages/api-client/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Public REST API of Honest Robin: Time. Authenticate with a personal access token (`Authorization: Bearer hrt_…`). Session-authenticated browser clients send the `HonestRobin-Account-Id` header to pick an account.",
    "license" : {
      "name" : "AGPL-3.0-only",
      "url" : "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "title" : "Honest Robin: Time API",
    "version" : "v1"
  },
  "security" : [ {
    "token" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/invoice_settings" : {
      "get" : {
        "operationId" : "invoice_settings_get",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceSettingsView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      },
      "patch" : {
        "operationId" : "invoice_settings_update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/InvoiceSettingsInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceSettingsView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
    "/api/v1/invoices" : {
      "get" : {
        "operationId" : "invoice_list",
        "parameters" : [ {
          "in" : "query",
          "name" : "state",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "client_id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageInvoiceSummaryView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      },
      "post" : {
        "operationId" : "invoice_create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/InvoiceInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
...
    "/api/v1/invoices/uninvoiced" : {
      "get" : {
        "operationId" : "invoice_uninvoiced",
        "parameters" : [ {
          "in" : "query",
          "name" : "client_id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UninvoicedPreview"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
    "/api/v1/invoices/{id}" : {
      "delete" : {
        "operationId" : "invoice_delete",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "invoices" ]
      },
      "get" : {
        "operationId" : "invoice_get",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      },
      "patch" : {
        "operationId" : "invoice_update",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/InvoiceInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
...
    "/api/v1/invoices/{id}/einvoice" : {
      "get" : {
        "operationId" : "invoice_einvoice",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "format",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
    "/api/v1/invoices/{id}/einvoice/readiness" : {
      "get" : {
        "operationId" : "invoice_einvoice_readiness",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EInvoiceReadiness"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
...
    "/api/v1/invoices/{id}/mark_sent" : {
      "post" : {
        "operationId" : "invoice_mark_sent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
    "/api/v1/invoices/{id}/payments" : {
      "post" : {
        "operationId" : "invoice_add_payment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PaymentInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
    "/api/v1/invoices/{id}/payments/{paymentId}" : {
      "delete" : {
        "operationId" : "invoice_delete_payment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "paymentId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
    "/api/v1/invoices/{id}/pdf" : {
      "get" : {
        "operationId" : "invoice_pdf",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
    "/api/v1/invoices/{id}/send" : {
      "get" : {
        "operationId" : "invoice_send_defaults",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendDefaults"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      },
      "post" : {
        "operationId" : "invoice_send",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendInvoiceInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
    "/api/v1/invoices/{id}/void" : {
      "post" : {
        "operationId" : "invoice_void",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "invoices" ]
      }
    },
...
    "/api/v1/public/invoices/{token}" : {
      "get" : {
        "operationId" : "public_invoice_get",
        "parameters" : [ {
          "in" : "path",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicInvoiceView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "public" ]
      }
    },
...
    "/api/v1/public/invoices/{token}/pdf" : {
      "get" : {
        "operationId" : "public_invoice_pdf",
        "parameters" : [ {
          "in" : "path",
          "name" : "token",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "public" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "InvoiceSettingsInput" : {
        "properties" : {
          "bcc_sender" : {
            "type" : "boolean"
          },
          "default_tax1_name" : {
            "type" : "string"
          },
          "default_tax1_percent" : {
            "type" : "number"
          },
          "default_tax2_name" : {
            "type" : "string"
          },
          "default_tax2_percent" : {
            "type" : "number"
          },
          "einvoice_hybrid_pdf" : {
            "type" : "boolean"
          },
          "footer" : {
            "type" : "string"
          },
          "invoice_contact_email" : {
            "type" : "string"
          },
          "invoice_contact_name" : {
            "type" : "string"
          },
          "invoice_contact_phone" : {
            "type" : "string"
          },
          "next_number" : {
            "type" : "integer",
            "format" : "int64"
          },
          "notes" : {
            "type" : "string"
          },
          "number_padding" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number_per_year" : {
            "type" : "boolean"
          },
          "number_prefix" : {
            "type" : "string"
          },
          "payment_instructions" : {
            "type" : "string"
          },
          "payment_terms_days" : {
            "type" : "integer",
            "format" : "int32"
          },
          "reminder_days" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "reminders_enabled" : {
            "type" : "boolean"
          }
        }
      },
      "InvoiceSettingsView" : {
        "properties" : {
          "bcc_sender" : {
            "type" : "boolean"
          },
          "default_tax1_name" : {
            "type" : "string"
          },
          "default_tax1_percent" : {
            "type" : "number"
          },
          "default_tax2_name" : {
            "type" : "string"
          },
          "default_tax2_percent" : {
            "type" : "number"
          },
          "einvoice_hybrid_pdf" : {
            "type" : "boolean"
          },
          "einvoice_hybrid_pdf_active" : {
            "type" : "boolean"
          },
          "footer" : {
            "type" : "string"
          },
          "invoice_contact_email" : {
            "type" : "string"
          },
          "invoice_contact_name" : {
            "type" : "string"
          },
          "invoice_contact_phone" : {
            "type" : "string"
          },
          "next_number" : {
            "type" : "integer",
            "format" : "int64"
          },
          "next_number_preview" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "number_padding" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number_per_year" : {
            "type" : "boolean"
          },
          "number_prefix" : {
            "type" : "string"
          },
          "payment_instructions" : {
            "type" : "string"
          },
          "payment_terms_days" : {
            "type" : "integer",
            "format" : "int32"
          },
          "reminder_days" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "reminders_enabled" : {
            "type" : "boolean"
          }
        },
        "required" : [ "bcc_sender", "einvoice_hybrid_pdf_active", "next_number", "next_number_preview", "number_padding", "number_per_year", "number_prefix", "payment_terms_days", "reminder_days", "reminders_enabled" ]
      },
...
      "PageInvoiceSummaryView" : {
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/InvoiceSummaryView"
            }
          },
          "next_cursor" : {
            "type" : "string"
          }
        },
        "required" : [ "data" ]
      },
...
      "Ref" : {
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ]
      },
...
      "UninvoicedPreview" : {
        "properties" : {
          "amount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "client" : {
            "$ref" : "#/components/schemas/Ref"
          },
          "currency" : {
            "type" : "string"
          },
          "expense_amount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "projects" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UninvoicedProject"
            }
          }
        },
        "required" : [ "amount", "client", "currency", "expense_amount", "projects" ]
      },
...
    },
    "securitySchemes" : {
      "token" : {
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}